refactor: Move classes to seperate files and add initial type hints
